% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{col_classes}
\alias{col_classes}
\title{Table column classes}
\usage{
col_classes(.tbl)
}
\description{
Returns a \code{\link{tibble}} containing the class of each column in a given
table. Works for in-memory and database tables.
}
\examples{

col_classes(day_metrics())
#> # A tibble: 2 x 7
#> id      source_name source_id metric_type metric_date value   created_at
#> <chr>   <chr>       <chr>     <chr>       <chr>       <chr>   <chr>
#> 1 integer integer     character integer     POSIXct     numeric character
#> 2 integer integer     character integer     POSIXt      numeric character

}
