% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_social_metrics.R
\name{diff_metrics}
\alias{diff_metrics}
\title{Differences cumulative metrics in `raw_social_metrics_cs`}
\usage{
diff_metrics(.tbl)
}
\description{
Some metric types in `raw_social_metrics_cs` are cumulative, others are not. This
function takes all cumulative metrics and transforms them into non-cumulative
by taking lag 1 differences of metrics with themselves.
}
\details{
Works only for in-memory tables. Must enrich source names and metric types
with `with_source_names` first.
}
