% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/right_holder_pseudos.R
\name{find_right_holder}
\alias{find_right_holder}
\title{Finds a right_holder ID by its pseudonymn.}
\usage{
find_right_holder(
  name,
  mode = c("any", "all", "main"),
  if_absent = c("raise_error", "return_NA")
)
}
\arguments{
\item{name}{a string representing the artist's name.}

\item{mode}{match mode. 'any' will return any matching pseudonym; 'all' will
return all pseudonyms, 'main' will return only the main pseudonym,
or error out if there is more than one.}

\item{if_absent}{how to behave when a pseudonym is missing. Can be either
'raise_error' (default), which will raise an error; or
'return_NA', which will return NAs for pseudonyms that
cannot be matched.}
}
\value{
an integer vector representing the right holder IDs associated with this
name.
}
\description{
Finds a right_holder ID by its pseudonymn.
}
