% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/right_holders.R
\name{find_right_holders}
\alias{find_right_holders}
\title{As \code{\link{find_right_holder}}, but accepts multiple right holder
pseudonyms.}
\usage{
find_right_holders(..., .dots = NULL, mode = c("any", "all", "main"))
}
\arguments{
\item{...}{a set of right holder names to resolve.}

\item{match}{mode. 'any' will return any matching pseudonym; 'all' will
return all pseudonyms, 'main' will return only the main pseudonym,
or error out if there is more than one.}
}
\value{
a \code{\link{data.frame}} containing a `pseudo` and a `right_holder_id`
        column. Depending on the match mode, may return multiple rows for
        a given pseudo.
}
\description{
As \code{\link{find_right_holder}}, but accepts multiple right holder
pseudonyms.
}
\seealso{
find_right_holder
}
