% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{for_dates}
\alias{for_dates}
\title{Filters a given table by date bracket}
\usage{
for_dates(.tbl, start, end = NULL)
}
\arguments{
\item{start}{the start of a date bracket.}

\item{end}{the end of a date bracket. Defaults to today.}
}
\description{
Will round start and end dates according to the granularity of the receiving
table. If end date is omitted, will default to the current date.
}
