% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{for_location}
\alias{for_location}
\title{Returns location-specific data. Supports returning data by city, state,
region, or country.}
\usage{
for_location(.tbl, ...)
}
\description{
If the name of an administrative region is ambiguous (e.g. the city of York
exists in both England and the US), will raise an error unless a disambiguating
outer administrative region (the country, in this case) is also supplied.
}
\details{
Not all tables support location filtering, and not all tables that support
location filtering support all location types. See \link{supported_location_types}
before attempting to query a table.
}
