% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{for_location_type}
\alias{for_location_type}
\title{In tables which support multiple location types, returns data specific to
one of them. See "supported_location_types".}
\usage{
for_location_type(.tbl, location_type)
}
\description{
In tables which support multiple location types, returns data specific to
one of them. See "supported_location_types".
}
