% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{for_metric_type}
\alias{for_metric_type}
\title{Filters a given table by metric type}
\usage{
for_metric_type(.tbl, metric_type, add_metric_types = TRUE)
}
\description{
Supported metric types depend on the table and, for some tables, may also
depend on the source. For that reason, it may not be possible to filter by
a metric type before specifying a source filter. See
\code{\link{supported_metric_types}} to verify which metric types are supported.
}
\seealso{
supported_metric_types
}
