% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{for_right_holder}
\alias{for_right_holder}
\title{Filters a table by (a set of) right holder(s)}
\usage{
for_right_holder(.tbl, ..., .dots = NULL)
}
\arguments{
\item{...}{a set of numeric IDs and/or string names of the desired right holders.
For each string supplied, will attempt translation with \code{\link{find_right_holder}}
first.}

\item{.dots}{optionally, specify the arguments as a list/vector under this parameter.}
}
\description{
Filters a table by right holder, translating its string name into the
corresponding numeric ID as needed. For some tables, it may be necessary to
add right holder information with \code{\link{with_right_holders}} first.
}
\examples{

week_metrics() \%>\%
   for_right_holder('Marília Mendonça')

day_metrics() \%>\%
   with_right_holders \%>\%
   for_right_holder('Marília Mendonça', 'MC Kevinho')

}
