% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{for_source}
\alias{for_source}
\title{Filters a given table by source (e.g. `youtube`, or `spotify`)}
\usage{
for_source(.tbl, ..., .dots = NULL)
}
\description{
Valid source names depend on the actual source. Use \code{\link{supported_sources}}
to find out which ones are supported.
}
