% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{is_in_memory}
\alias{is_in_memory}
\title{Checks if a table is in memory}
\usage{
is_in_memory(.tbl)
}
\value{
\code{True} if the table is in memory, or \code{False} otherwise.
}
\description{
Checks if a table reference points to a table in memory or to a remote table
in the database.
}
