% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locations.R
\docType{data}
\name{LOCATION_TYPES}
\alias{LOCATION_TYPES}
\title{The list of possible location types supported by tables which support location
filtering. Not all tables support all location types (see \link{supported_location_types}).}
\format{
An object of class \code{character} of length 4.
}
\usage{
LOCATION_TYPES
}
\description{
The list of possible location types supported by tables which support location
filtering. Not all tables support all location types (see \link{supported_location_types}).
}
\keyword{datasets}
