% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_social_metrics.R
\docType{data}
\name{MAPPINGS}
\alias{MAPPINGS}
\title{Source-specific metric mappings}
\format{
An object of class \code{list} of length 5.
}
\usage{
MAPPINGS
}
\description{
List containing source-specific metrics with their original naming.
Some of those get mapped into \code{\link{STANDARD_METRICS}} through
the \code{\link{metric_type_mapping}} table.
}
\keyword{datasets}
