% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_metrics.R
\name{match_metrics}
\alias{match_metrics}
\title{Resolve metric type indices}
\usage{
match_metrics(metric_types, available_metrics = names(STANDARD_METRICS))
}
\description{
Matches a character vector of metric types to their corresponding metric
indices, raising an error for unknown metric types.
}
