% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\docType{data}
\name{MIN_DATE}
\alias{MIN_DATE}
\title{The smallest possible date}
\format{
An object of class \code{character} of length 1.
}
\usage{
MIN_DATE
}
\description{
The smallest possible date
}
\keyword{datasets}
