% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/right_holder_pseudos.R
\docType{data}
\name{PSEUDONYM_LOOKUP_STRATEGIES}
\alias{PSEUDONYM_LOOKUP_STRATEGIES}
\title{Pre-packaged strategies which might be supplied to find_right_holder(s) to
increase chances of a match success by transforming the input.}
\format{
An object of class \code{list} of length 1.
}
\usage{
PSEUDONYM_LOOKUP_STRATEGIES
}
\description{
Pre-packaged strategies which might be supplied to find_right_holder(s) to
increase chances of a match success by transforming the input.
}
\keyword{datasets}
