% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_social_metrics.R
\name{raw_social_metrics}
\alias{raw_social_metrics}
\alias{city_social_metrics}
\alias{state_social_metrics}
\alias{region_social_metrics}
\title{Global and regionalized per-artist, "raw" metrics}
\usage{
raw_social_metrics(scope = c("global", "city", "state", "region"))

city_social_metrics()

state_social_metrics()

region_social_metrics()
}
\arguments{
\item{scope}{a regionalized scope for the data. Valid scopes are `global`
(data for the whole planet or the whole country, default),
`city` (data per city), `state` (date per state), and
`region` (data per region).}
}
\description{
`*_social_metrics` are a set of per-artist tables which store "raw", daily
data from the various sources ingested by Playax. These are "raw" in the
sense that no processing is done on the data, and it is stored as-is.
}
\details{
Regionalized variants are avaiable by supplying a different _scope_ parameter,
or by calling `{city,state,region}_social_metrics` directly.
}
