% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/right_holder_external_ids.R
\name{right_holder_external_ids}
\alias{right_holder_external_ids}
\title{External IDs for right holders. Those will usually correspond to the IDs
that a right holder may take in various platforms (YouTube, Instagram, etc).}
\usage{
right_holder_external_ids()
}
\description{
External IDs for right holders. Those will usually correspond to the IDs
that a right holder may take in various platforms (YouTube, Instagram, etc).
}
