% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/right_holders.R
\name{right_holders}
\alias{right_holders}
\title{Returns the table of all right holders, with information such
as their name, genre, and language.}
\usage{
right_holders()
}
\description{
Returns the table of all right holders, with information such
as their name, genre, and language.
}
