% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\docType{data}
\name{STANDARD_METRICS}
\alias{STANDARD_METRICS}
\title{Standard metric types}
\format{
An object of class \code{list} of length 4.
}
\usage{
STANDARD_METRICS
}
\description{
When supported, standard metrics have a well-defined semantics. Usually
used with \code{\link{for_metric_type}}.
}
\keyword{datasets}
