% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{supported_location_types}
\alias{supported_location_types}
\title{Returns a character vector containing the location types supported by this
table. It should be a subset of \link{LOCATION_TYPES}.}
\usage{
supported_location_types(.tbl)
}
\description{
Returns a character vector containing the location types supported by this
table. It should be a subset of \link{LOCATION_TYPES}.
}
