% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{supported_metric_types}
\alias{supported_metric_types}
\title{Returns a character vector with supported metric types}
\usage{
supported_metric_types(.tbl, source, standard = TRUE)
}
\arguments{
\item{standard}{returns only standard metrics. Defaults to TRUE.}
}
\description{
These values can be later used with \code{\link{for_metric_type}}, with the
caveat that source selection with \code{\link{for_source}} may be required
first.
}
