% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{supported_sources}
\alias{supported_sources}
\title{Returns a character vector with supported sources}
\usage{
supported_sources(.tbl)
}
\description{
These are values can be used with \code{\link{for_source}}.
}
