% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracks.R
\name{tracks_for}
\alias{tracks_for}
\title{Lists tracks for a right holder}
\usage{
tracks_for(right_holder)
}
\description{
Given a numeric right holder id or string pseudonym, returns all tracks that
have been released by that right holder.
}
\examples{

  tracks_for('Ludmilla')

}
