% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/right_holders.R
\name{with_genres}
\alias{with_genres}
\title{Enriches a table containing right holder information column with root
genres, placed under a `root_genre` column.}
\usage{
with_genres(.tbl)
}
\description{
Enriches a table containing right holder information column with root
genres, placed under a `root_genre` column.
}
