% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/right_holders.R
\name{with_language}
\alias{with_language}
\title{Enriches a table containing right holder information with a "language"
column, representing the artist's main language.}
\usage{
with_language(.tbl)
}
\description{
Enriches a table containing right holder information with a "language"
column, representing the artist's main language.
}
