% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{with_metric_types}
\alias{with_metric_types}
\title{Adds symbolic metric types into the current table, for all sources}
\usage{
with_metric_types(.tbl)
}
\description{
Decodes metric types in the current table into strings, either by replacing
a column in the original table or by creating a new one. May have to copy
table contents in-memory if user has no write access to the underlying
database.
}
