% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/right_holder_pseudos.R
\name{with_pseudos}
\alias{with_pseudos}
\title{Enriches table with right holder pseudonyms}
\usage{
with_pseudos(.tbl, mode = c("any", "main", "all"))
}
\arguments{
\item{mode}{the pseudonyms to join with. See \code{\link{right_holder_pseudos}}
for mode details.}
}
\description{
Enriches the current table with data from the \code{\link{right_holder_pseudos}}
table, plugging in pseudonyms.
}
