% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/right_holders.R
\name{with_right_holders}
\alias{with_right_holders}
\title{Enriches the current table with right holder information. By default,
drops all rows that cannot be matched to any right holder, or which are
associated to invalid right holders.}
\usage{
with_right_holders(.tbl, if_missing = NULL, drop_invalid = TRUE)
}
\arguments{
\item{if_missing}{(deprecated) enriches the table only if a certain column
is absent. Defaults to `NULL`.}

\item{drop_invalid}{drops rows which cannot be matched to any right holder,
and right holders which are declared as not valid. Defaults
to TRUE.}
}
\description{
Enriches the current table with right holder information. By default,
drops all rows that cannot be matched to any right holder, or which are
associated to invalid right holders.
}
