% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{with_source_names}
\alias{with_source_names}
\title{Adds symbolic source names into the current table}
\usage{
with_source_names(.tbl)
}
\description{
Decodes source names in the current table into strings, either by replacing
a column in the original table or by creating a new one. May have to copy
table contents in-memory if user has no write access to the underlying
database.
}
