% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracks.R
\name{with_tracks}
\alias{with_tracks}
\title{Adds track info to the current table, if applicable.}
\usage{
with_tracks(.tbl, drop_invalid = TRUE)
}
\arguments{
\item{drop_invalid}{if set to TRUE (default), drops rows which have
no corresponding track. Otherwise keeps them.}
}
\description{
For tables which carry a `track_id`, adds info from the tracks table.
}
