% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/right_holders.R
\name{right_holders}
\alias{right_holders}
\title{Right holder base data}
\usage{
right_holders()
}
\description{
Table containing all right holders, with information such as
their name, genre, and language.
}
