% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/right_holders.R
\name{with_pseudos}
\alias{with_pseudos}
\title{Enriches a table containing right holder information with associated
pseudonyms.}
\usage{
with_pseudos(.tbl, main = TRUE)
}
\arguments{
\item{main}{for right holders having more than one pseudonym, returns
only the one registered as its "main" pseudonym, if any. Note that data
errors mean that it may happen that an artist has multiple "main"
pseudonyms.}
}
\description{
Enriches a table containing right holder information with associated
pseudonyms.
}
